requirements.txt
asgiref==3.5.2
Django==4.1.2
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
Pillow==9.3.0
psycopg2-binary==2.9.4
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.5
